pyproject.toml
[project]
name = "amy"
version = "0.1.0"
description = "Personal AI Agent based on LangChain"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.2.0",
    "pyyaml>=6.0",
    "structlog>=24.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
agentlightning = [
    "agentlightning[apo]>=0.1.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]
